# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Wegweiser API
  version: 0.1.0
  description: |
    The API every Wegweiser client speaks. The web interface and the `weg`
    command line are both clients of it, and neither of them reaches the
    database any other way (architecture invariant 1).

    This document is the source of truth. Clients are generated from it, so an
    endpoint that is not written down here does not exist.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /api/v1
security:
  - bearerToken: []
...
paths:
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List the tokens that may use this API
      description: |
        Revoked and expired tokens are listed too, so that a name in the
        history can still be looked up. No response ever carries a secret: what
        is stored is its SHA-256, and the leading characters kept in the clear
        are there to tell two tokens apart, not to authenticate with.
      responses:
        "200":
          description: Every token.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Token" }
        default: { $ref: "#/components/responses/Problem" }

    post:
      operationId: createToken
      tags: [tokens]
      summary: Mint a token
      description: |
        The secret comes back in this response and never again. It is 256 bits
        from a cryptographic source; what the server keeps is its SHA-256, so a
        copy of the database is not a copy of the credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateToken" }
      responses:
        "201":
          description: The token, with its secret shown for the only time.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenCreated" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: |
        A named API token. Tokens carry the scopes `read`, `write` and `admin`;
        an endpoint that changes anything needs `write`.
...
  responses:
    Problem:
      description: The request could not be carried out.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: |
        A failure, in the shape of RFC 9457. Anything a client has to act on
        travels as a field rather than as prose in `detail`, because no client
        should be parsing English.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: A URI reference naming the kind of failure.
        title:
          type: string
          description: A short, stable summary of the kind of failure.
        status:
          type: integer
          description: The HTTP status code, repeated for clients that lost it.
        detail:
          type: string
          description: What went wrong with this particular request.
        instance:
          type: string
          description: The request this failure belongs to.
        conflicts:
          type: array
          description: Reverse entries that were not written because another name already claims the address.
          items: { $ref: "#/components/schemas/Conflict" }
        missingZones:
          type: array
          description: Reverse zones that would be needed and do not exist.
          items: { $ref: "#/components/schemas/MissingZone" }
    ReverseConflictPolicy:
      type: string
      enum: [first-wins, last-wins, multi, reject]
      description: |
        What happens when an address already answers a reverse lookup with some
        other name (docs/decisions/ D3).

        - `first-wins` keeps the entry that is there and reports the conflict.
          The default, and the only value that never changes an answer nobody
          asked to change.
        - `last-wins` replaces a generated entry with the new one, still
          reporting the conflict. An entry somebody wrote by hand is never
          replaced.
        - `multi` keeps both, which is the literal reading of "generate the
          reverse entry" and turns a routine change into a multi-record PTR set
          that reverse-lookup checks are not built for.
        - `reject` refuses the whole change.
...
    Scope:
      type: string
      description: |
        What a token may do. The scopes are ordered (admin allows everything
        write allows, and write everything read allows) so a token carries the
        strongest one it needs rather than a list.
      enum: [read, write, admin]
    Token:
      type: object
      description: A credential, without the secret it authenticates with.
      required: [id, name, prefix, scopes, createdAt]
      properties:
        id: { type: string }
        name:
          type: string
          description: What this token is for, as a person would say it.
        prefix:
          type: string
          description: |
            The leading characters of the secret, kept in the clear so that two
            tokens can be told apart in a list and in the history.
        scopes:
          type: array
          items: { $ref: "#/components/schemas/Scope" }
        createdAt: { type: string, format: date-time }
        lastUsedAt:
          type: string
          format: date-time
          description: Absent for a token that has never authenticated a request. Advisory, and may lag.
        expiresAt:
          type: string
          format: date-time
          description: Absent for a token that does not expire.
        revokedAt:
          type: string
          format: date-time
          description: Absent for a token that is still usable.
    CreateToken:
      type: object
      required: [name, scopes]
      properties:
        name: { type: string }
        scopes:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/Scope" }
        expiresAt:
          type: string
          format: date-time
          description: When the token stops working. Absent means it does not.
    TokenCreated:
      type: object
      required: [token, secret]
      properties:
        token: { $ref: "#/components/schemas/Token" }
        secret:
          type: string
          description: |
            The credential itself, shown here and nowhere else. There is no way
            to read it again, because the server does not have it.
...
    Conflict:
      type: object
      description: |
        An address already answers with another name (docs/decisions/ D3). It
        is data rather than an error: which name should be canonical is a
        decision for a person.
      required: [address, existingName, requestedName, policy]
      properties:
        address: { type: string }
        existingName: { type: string }
        requestedName: { type: string }
        existingRecordId: { type: string }
        policy:
          allOf:
            - $ref: "#/components/schemas/ReverseConflictPolicy"
          description: |
            The policy that decided what happened here, so a client can say
            whether the existing name was kept, replaced or joined rather than
            leaving the reader to guess from the records.
    MissingZone:
      type: object
      description: |
        A reverse entry has nowhere to go, because no zone covers the address.
        Creating one is an assertion of authority, so it is offered rather than
        done (docs/decisions/ D6).
      required: [address, zoneName]
      properties:
        address: { type: string }
        zoneName:
          type: string
          description: The reverse zone that would be needed.